+++ includes/class-plugin.php
@@ -51,8 +51,9 @@ class Alynt_Drime_Backups_Dashboard_Plugin {
 		$sites           = new Alynt_Drime_Backups_Dashboard_Site_Repository();
 		$this->snapshots = new Alynt_Drime_Backups_Dashboard_Snapshot_Repository();
 		$classifier      = new Alynt_Drime_Backups_Dashboard_Status_Classifier();
-		$this->poller    = new Alynt_Drime_Backups_Dashboard_Poller( $sites, $this->snapshots, $classifier );
-		$diagnostics     = new Alynt_Drime_Backups_Dashboard_Diagnostics( $sites, $this->snapshots, $classifier );
+		$event_log       = new Alynt_Drime_Backups_Dashboard_Event_Log();
+		$this->poller    = new Alynt_Drime_Backups_Dashboard_Poller( $sites, $this->snapshots, $classifier, null, null, null, null, $event_log );
+		$diagnostics     = new Alynt_Drime_Backups_Dashboard_Diagnostics( $sites, $this->snapshots, $classifier, $event_log );
 
 		$this->admin_page                 = new Alynt_Drime_Backups_Dashboard_Admin_Page(
 			$sites,
@@ -62,7 +63,7 @@ class Alynt_Drime_Backups_Dashboard_Plugin {
 			$this->poller,
 			$diagnostics
 		);
-		$this->enrollment_rest_controller = new Alynt_Drime_Backups_Dashboard_Enrollment_REST_Controller();
+		$this->enrollment_rest_controller = new Alynt_Drime_Backups_Dashboard_Enrollment_REST_Controller( null, null, null, $event_log );
 
 		$this->hooks();
 	}